feat: grouped listings/carousel groups/permissions admin pages, cloud logs, offers + wide fixes
Batches together several in-flight threads from this session and prior
concurrent sessions; syncs the appkit submodule pointer.

- New admin pages: grouped-listings, permissions (read-only catalog), and
  RSC detail pages for item-requests/moderation/orders/reports/scammers/
  support-tickets ([id] routes, previously list-only)
- New API routes: /api/admin/carousels(+[id]) for carousel groups,
  /api/admin/maintenance/cloud-logs backing the new maintenance view
- New audits wired into run-audits.mjs + the Stop hook: audit-listing-filter-parity
  (Root Cause #30 — SSR/client default-filter divergence on public listing
  pages), audit-nav-page-wiring (Root Cause #29 — dead nav links / orphan
  pages), audit-select-wrapper-classname, audit-tester-checklist-hrefs,
  audit-filter-tab-enums
- Root cause fix: /api/products defaults status to published when the caller
  sends none, matching /api/events/reviews/stores — closes a gap where every
  client-driven refetch on 12 public listing surfaces could drop the
  published-only guard
- CLAUDE.md: Root Cause #30 entry, corrected Media Filename Slug Patterns
  table (verified against id-generators.ts — most rows previously didn't
  match real generator output), offer- slug prefix documented
- codebaseexports.md: reconciled against the new/removed files in this batch
  (audit scripts, carousel groups, cloud logs, SiteMark, admin-permissions
  mirror, templates feature removal)
- Removed store product-templates feature (SellerTemplatesView, page shims,
  API routes) — repository/schema layer intentionally left in place
- FAQ page client component, watermark resolution lib, offer actions

+++ src/actions/offer.actions.ts
@@ -37,14 +37,14 @@ import { ERR_RATE_LIMIT } from "./_constants";
 
 const makeOfferSchema = z.object({
   productId: z.string().min(1),
-  offerAmount: z.number().positive().int(),
+  offerAmount: z.number().positive(),
   buyerNote: z.string().max(300).optional(),
 });
 
 const respondToOfferSchema = z.object({
   offerId: z.string().min(1),
   action: z.enum(["accept", "decline", "counter"]),
-  counterAmount: z.number().positive().int().optional(),
+  counterAmount: z.number().positive().optional(),
   sellerNote: z.string().max(300).optional(),
 });
 
@@ -58,7 +58,7 @@ const withdrawOfferSchema = z.object({
 
 const buyerCounterSchema = z.object({
   offerId: z.string().min(1),
-  counterAmount: z.number().positive().int(),
+  counterAmount: z.number().positive(),
   buyerNote: z.string().max(300).optional(),
 });
